resin: don't die immediately when inferior goes away
We want to give the Erlang peer a bit of time to receive our
message about the status of the inferior process. We can shutdown
the inferior-handler threads, but leave the main port I/O loop going.

If Erlang is still alive it will close this port, and when that happens
we can gracefully terminate. In the future we might consider adding some
sort of timeout here so that we don't wait forever for a dead peer.