(bump version) v0.2.1
This release includes the following changes:

      (style) fix whitespace errors in resin
      (docs) update readme to reflect v0.2
      (docs) update LICENSE copyright year
      (resin) do not use buffered I/O

This release includes one fairly major change to `resin`, which is that for
communication between the `lacca` and `resin` we now use the actual file
descriptor instead of going through the stdlibs stdin/stdout primitives.

The Erlang/OTP manual instructs that port authors do not use buffered
I/O when communicating w/ an external program. This is our attempt to
comply with that requirement. Coincidentally backporting this to v0.1  fixes
the issue in which CBOR messages were getting clobbered on the wire.